+++ app/Providers/RepositoryLayerProvider.php
@@ -4,7 +4,9 @@ namespace App\Providers;
 
 use App\Repositories\CategoryRepository;
 use App\Repositories\Interfaces\ICategoryRepository;
+use App\Repositories\Interfaces\IItemRepository;
 use App\Repositories\Interfaces\IListRepository;
+use App\Repositories\ItemRepository;
 use App\Repositories\ListRepository;
 use Illuminate\Support\ServiceProvider;
 
@@ -17,6 +19,7 @@ class RepositoryLayerProvider extends ServiceProvider
     {
         $this->app->bind(ICategoryRepository::class, CategoryRepository::class);
         $this->app->bind(IListRepository::class, ListRepository::class);
+        $this->app->bind(IItemRepository::class, ItemRepository::class);
     }
 
     /**