+++ app/Providers/ServiceLayerProvider.php
@@ -4,7 +4,9 @@ namespace App\Providers;
 
 use App\Services\CategoryService;
 use App\Services\Interfaces\ICategoryService;
+use App\Services\Interfaces\IItemService;
 use App\Services\Interfaces\IListService;
+use App\Services\ItemService;
 use App\Services\ListService;
 use Illuminate\Support\ServiceProvider;
 
@@ -17,6 +19,7 @@ class ServiceLayerProvider extends ServiceProvider
     {
         $this->app->bind(ICategoryService::class, CategoryService::class);
         $this->app->bind(IListService::class, ListService::class);
+        $this->app->bind(IItemService::class, ItemService::class);
     }
 
     /**