+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\ItemController;
 use App\Http\Controllers\ListController;
 use Illuminate\Support\Facades\Route;
 
@@ -19,14 +20,12 @@ Route::group([
 //endregion
 
 //region Category
-Route::get('/category/{id?}', [CategoryController::class, 'index']);
-
 Route::group([
     'middleware' => 'api',
     'prefix' => 'category'
 ], function ($router) {
-
-    Route::put('/', [CategoryController::class, 'store']);
+    Route::get('/category/{id?}', [CategoryController::class, 'index']);
+    Route::post('/', [CategoryController::class, 'store']);
     Route::delete('/', [CategoryController::class, 'destroy']);
 });
 
@@ -38,7 +37,19 @@ Route::group([
     'prefix' => 'lists'
 ], function ($router) {
     Route::get('/{id?}', [ListController::class, 'index']);
-    Route::put('/', [ListController::class, 'store']);
+    Route::post('/', [ListController::class, 'store']);
     Route::delete('/{id}', [ListController::class, 'destroy']);
+});
+//endregion
+
+//region Item
+Route::group([
+    'middleware' => 'api',
+    'prefix' => 'items'
+], function ($router) {
+    Route::get('/{listId?}/{itemId?}', [ItemController::class, 'index']);
+    Route::post('/', [ItemController::class, 'store']);
+    Route::patch('/{id}', [ItemController::class, 'update']);
+    Route::delete('/', [ItemController::class, 'destroy']);
 });
 //endregion